+++ flink-ml-parent/flink-ml-lib/src/test/java/org/apache/flink/ml/common/utils/DataStreamConversionUtilTest.java
@@ -21,9 +21,9 @@ package org.apache.flink.ml.common.utils;
 
 import org.apache.flink.api.common.functions.MapFunction;
 import org.apache.flink.api.common.typeinfo.TypeInformation;
-import org.apache.flink.api.java.DataSet;
-import org.apache.flink.api.java.ExecutionEnvironment;
 import org.apache.flink.ml.common.MLEnvironmentFactory;
+import org.apache.flink.streaming.api.datastream.DataStream;
+import org.apache.flink.streaming.api.environment.StreamExecutionEnvironment;
 import org.apache.flink.table.api.Table;
 import org.apache.flink.table.api.TableSchema;
 import org.apache.flink.table.api.ValidationException;
@@ -34,22 +34,28 @@ import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
 
-import java.util.Collections;
-import java.util.List;
-
 /**
- * Unit test for RowTypeDataSet.
+ * Unit Test for DataStreamConversionUtil.
  */
-public class RowTypeDataSetTest {
+public class DataStreamConversionUtilTest {
 	@Rule
 	public ExpectedException thrown = ExpectedException.none();
 
 	@Test
-	public void testForceTypeSchema() {
-		ExecutionEnvironment env = MLEnvironmentFactory.getDefault().getExecutionEnvironment();
+	public void test() throws Exception {
+		StreamExecutionEnvironment env = MLEnvironmentFactory.getDefault().getStreamExecutionEnvironment();
 
-		DataSet<Row> input = env.fromElements(Row.of("s1")).map(new GenericTypeMap());
-		Table table2 = DataSetConversionUtil.toTable(MLEnvironmentFactory.DEFAULT_ML_ENVIRONMENT_ID,
+		DataStream<Row> input = env.fromElements(Row.of("a"));
+
+		Table table1 = DataStreamConversionUtil.toTable(MLEnvironmentFactory.DEFAULT_ML_ENVIRONMENT_ID, input, new String[]{"word"});
+		Assert.assertEquals(
+			new TableSchema(new String[]{"word"}, new TypeInformation[]{TypeInformation.of(String.class)}),
+			table1.getSchema()
+		);
+
+		input = input.map(new GenericTypeMap());
+		Table table2 = DataStreamConversionUtil.toTable(
+			MLEnvironmentFactory.DEFAULT_ML_ENVIRONMENT_ID,
 			input,
 			new String[]{"word"},
 			new TypeInformation[]{TypeInformation.of(Integer.class)}
@@ -58,29 +64,15 @@ public class RowTypeDataSetTest {
 			new TableSchema(new String[]{"word"}, new TypeInformation[]{TypeInformation.of(Integer.class)}),
 			table2.getSchema()
 		);
-	}
 
-	@Test
-	public void testExceptionWithoutTypeSchema() {
 		thrown.expect(ValidationException.class);
-		ExecutionEnvironment env = MLEnvironmentFactory.getDefault().getExecutionEnvironment();
-		DataSet<Row> input = env.fromElements(Row.of("s1")).map(new GenericTypeMap());
-		DataSetConversionUtil.toTable(MLEnvironmentFactory.DEFAULT_ML_ENVIRONMENT_ID, input, new String[]{"f0"});
-	}
+		DataStreamConversionUtil.toTable(MLEnvironmentFactory.DEFAULT_ML_ENVIRONMENT_ID, input, new String[]{"f0"});
 
-	@Test
-	public void testBasicConvert() throws Exception {
-		ExecutionEnvironment env = MLEnvironmentFactory.getDefault().getExecutionEnvironment();
+		DataStream<Row> output = DataStreamConversionUtil.fromTable(MLEnvironmentFactory.DEFAULT_ML_ENVIRONMENT_ID, table1);
 
-		DataSet<Row> input = env.fromElements(Row.of("a"));
+		output.print();
 
-		Table table1 = DataSetConversionUtil.toTable(MLEnvironmentFactory.DEFAULT_ML_ENVIRONMENT_ID, input, new String[]{"word"});
-		Assert.assertEquals(
-			new TableSchema(new String[]{"word"}, new TypeInformation[]{TypeInformation.of(String.class)}),
-			table1.getSchema()
-		);
-		List<Row> list = DataSetConversionUtil.fromTable(MLEnvironmentFactory.DEFAULT_ML_ENVIRONMENT_ID, table1).collect();
-		Assert.assertEquals(Collections.singletonList(Row.of("a")), list);
+		env.execute();
 	}
 
 	private static class GenericTypeMap implements MapFunction<Row, Row> {
